#112 - Refactor namespaces + remove dead wood
