+++ src/static/functions.py
